Revert "Introduce executeQuery on Driver level (#1321)"
This reverts commit 477577bba10d91c3969837e4288e3a802a05c1c3.
