+++ driver/src/main/java/org/neo4j/driver/internal/DriverFactory.java
@@ -260,8 +260,7 @@ public class DriverFactory {
      */
     protected InternalDriver createDriver(
             SecurityPlan securityPlan, SessionFactory sessionFactory, MetricsProvider metricsProvider, Config config) {
-        return new InternalDriver(
-                config.queryBookmarkManager(), securityPlan, sessionFactory, metricsProvider, config.logging());
+        return new InternalDriver(securityPlan, sessionFactory, metricsProvider, config.logging());
     }
 
     /**
