--- driver/src/main/java/org/neo4j/driver/internal/EagerResultTransformer.java
@@ -1,34 +1,0 @@
-/*
- * Copyright (c) "Neo4j"
- * Neo4j Sweden AB [http://neo4j.com]
- *
- * This file is part of Neo4j.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.neo4j.driver.internal;
-
-import java.io.Serial;
-import org.neo4j.driver.EagerResult;
-import org.neo4j.driver.Result;
-import org.neo4j.driver.ResultTransformer;
-
-public final class EagerResultTransformer implements ResultTransformer<EagerResult> {
-    @Serial
-    private static final long serialVersionUID = 67912015934658230L;
-
-    @Override
-    public EagerResult transform(Result result) {
-        return new EagerResultValue(result.keys(), result.list(), result.consume());
-    }
-}
