+++ driver/src/main/java/org/neo4j/driver/internal/SessionFactoryImpl.java
@@ -61,7 +61,7 @@ public class SessionFactoryImpl implements SessionFactory {
                 parseFetchSize(sessionConfig),
                 sessionConfig.impersonatedUser().orElse(null),
                 logging,
-                sessionConfig.bookmarkManager().orElse(NoOpBookmarkManager.INSTANCE));
+                sessionConfig.bookmarkManager().orElse(new NoOpBookmarkManager()));
     }
 
     private Set<Bookmark> toDistinctSet(Iterable<Bookmark> bookmarks) {
