+++ driver/src/test/java/org/neo4j/driver/internal/InternalDriverTest.java
@@ -18,17 +18,12 @@
  */
 package org.neo4j.driver.internal;
 
+import static org.junit.Assert.assertTrue;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.BDDMockito.given;
-import static org.mockito.BDDMockito.then;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.spy;
-import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static org.neo4j.driver.internal.logging.DevNullLogging.DEV_NULL_LOGGING;
@@ -38,20 +33,8 @@ import static org.neo4j.driver.testutil.TestUtil.await;
 
 import java.util.concurrent.CompletableFuture;
 import org.junit.jupiter.api.Test;
-import org.mockito.ArgumentCaptor;
-import org.neo4j.driver.BookmarkManager;
-import org.neo4j.driver.BookmarkManagerConfig;
-import org.neo4j.driver.BookmarkManagers;
 import org.neo4j.driver.Config;
 import org.neo4j.driver.Metrics;
-import org.neo4j.driver.Query;
-import org.neo4j.driver.QueryConfig;
-import org.neo4j.driver.Result;
-import org.neo4j.driver.ResultTransformer;
-import org.neo4j.driver.Session;
-import org.neo4j.driver.SessionConfig;
-import org.neo4j.driver.TransactionCallback;
-import org.neo4j.driver.TransactionContext;
 import org.neo4j.driver.exceptions.ClientException;
 import org.neo4j.driver.exceptions.ServiceUnavailableException;
 import org.neo4j.driver.internal.metrics.DevNullMetricsProvider;
@@ -60,71 +43,6 @@ import org.neo4j.driver.internal.security.SecurityPlanImpl;
 import org.neo4j.driver.internal.util.Clock;
 
 class InternalDriverTest {
-    @Test
-    @SuppressWarnings("unchecked")
-    void shouldExecuteQuery() {
-        var sessionFactory = sessionFactoryMock();
-        var driver = spy(newDriver(sessionFactory));
-        var query = new Query("RETURN 1");
-        ResultTransformer<Object> resultTransformer = mock(ResultTransformer.class);
-        var bookmarkManager = mock(BookmarkManager.class);
-        var config = QueryConfig.builder(resultTransformer)
-                .withDatabase("database")
-                .withImpersonatedUser("user")
-                .withBookmarkManager(bookmarkManager)
-                .build();
-        var session = mock(Session.class);
-        given(driver.session(any(SessionConfig.class))).willReturn(session);
-        var result = mock(Result.class);
-        var txContext = mock(TransactionContext.class);
-        var expectedResult = new Object();
-        given(txContext.run(any(Query.class))).willReturn(result);
-        given(resultTransformer.transform(result)).willReturn(expectedResult);
-        given(session.executeWrite(any(TransactionCallback.class))).willAnswer(invocation -> {
-            TransactionCallback<Object> callback = invocation.getArgument(0);
-            return callback.execute(txContext);
-        });
-
-        var actualResult = driver.executeQuery(query, config);
-
-        var argument = ArgumentCaptor.forClass(SessionConfig.class);
-        then(driver).should(times(2)).session(argument.capture());
-        var sessionConfig = argument.getValue();
-        assertEquals(config.database(), sessionConfig.database());
-        assertEquals(config.bookmarkManager(mock(BookmarkManager.class)), sessionConfig.bookmarkManager());
-        then(session).should().executeWrite(any(TransactionCallback.class));
-        assertEquals(expectedResult, actualResult);
-    }
-
-    @Test
-    @SuppressWarnings("unchecked")
-    void shouldNotAllowReturningResultInExecuteQuery() {
-        var sessionFactory = sessionFactoryMock();
-        var driver = spy(newDriver(sessionFactory));
-        var query = new Query("RETURN 1");
-        ResultTransformer<Object> resultTransformer = mock(ResultTransformer.class);
-        var bookmarkManager = mock(BookmarkManager.class);
-        var config = QueryConfig.builder(resultTransformer)
-                .withDatabase("database")
-                .withImpersonatedUser("user")
-                .withBookmarkManager(bookmarkManager)
-                .build();
-        var session = mock(Session.class);
-        given(driver.session(any(SessionConfig.class))).willReturn(session);
-        var result = mock(Result.class);
-        var txContext = mock(TransactionContext.class);
-        given(txContext.run(any(Query.class))).willReturn(result);
-        given(resultTransformer.transform(result)).willReturn(result);
-        given(session.executeWrite(any(TransactionCallback.class))).willAnswer(invocation -> {
-            TransactionCallback<Object> callback = invocation.getArgument(0);
-            return callback.execute(txContext);
-        });
-
-        var exception = assertThrows(IllegalStateException.class, () -> driver.executeQuery(query, config));
-
-        assertTrue(exception.getMessage().contains("Illegal result returned"));
-    }
-
     @Test
     void shouldCloseSessionFactory() {
         SessionFactory sessionFactory = sessionFactoryMock();
@@ -195,11 +113,7 @@ class InternalDriverTest {
 
     private static InternalDriver newDriver(SessionFactory sessionFactory) {
         return new InternalDriver(
-                BookmarkManagers.defaultManager(BookmarkManagerConfig.builder().build()),
-                SecurityPlanImpl.insecure(),
-                sessionFactory,
-                DevNullMetricsProvider.INSTANCE,
-                DEV_NULL_LOGGING);
+                SecurityPlanImpl.insecure(), sessionFactory, DevNullMetricsProvider.INSTANCE, DEV_NULL_LOGGING);
     }
 
     private static SessionFactory sessionFactoryMock() {
@@ -216,11 +130,6 @@ class InternalDriverTest {
         }
 
         MetricsProvider metricsProvider = DriverFactory.getOrCreateMetricsProvider(config, Clock.SYSTEM);
-        return new InternalDriver(
-                BookmarkManagers.defaultManager(BookmarkManagerConfig.builder().build()),
-                SecurityPlanImpl.insecure(),
-                sessionFactory,
-                metricsProvider,
-                DEV_NULL_LOGGING);
+        return new InternalDriver(SecurityPlanImpl.insecure(), sessionFactory, metricsProvider, DEV_NULL_LOGGING);
     }
 }
