+++ testkit-backend/src/main/java/neo4j/org/testkit/backend/messages/requests/TestkitRequest.java
@@ -61,8 +61,7 @@ import reactor.core.publisher.Mono;
     @JsonSubTypes.Type(BookmarksSupplierCompleted.class),
     @JsonSubTypes.Type(BookmarksConsumerCompleted.class),
     @JsonSubTypes.Type(NewBookmarkManager.class),
-    @JsonSubTypes.Type(BookmarkManagerClose.class),
-    @JsonSubTypes.Type(ExecuteQuery.class)
+    @JsonSubTypes.Type(BookmarkManagerClose.class)
 })
 public interface TestkitRequest {
     TestkitResponse process(TestkitState testkitState);
