--- testkit-backend/src/main/java/neo4j/org/testkit/backend/messages/responses/EagerResult.java
@@ -1,42 +1,0 @@
-/*
- * Copyright (c) "Neo4j"
- * Neo4j Sweden AB [http://neo4j.com]
- *
- * This file is part of Neo4j.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package neo4j.org.testkit.backend.messages.responses;
-
-import java.util.List;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class EagerResult implements TestkitResponse {
-    private EagerResultBody data;
-
-    @Override
-    public String testkitName() {
-        return "EagerResult";
-    }
-
-    @Getter
-    @Builder
-    public static class EagerResultBody {
-        private final List<String> keys;
-        private final List<Record.RecordBody> records;
-        private final Summary.SummaryBody summary;
-    }
-}
